bp_function add sigmoid

+++ BP_multi-layer_feedforward_neural/Common.c
@@ -1,11 +1,11 @@
 #include "Common.h"
 
-float act_squashing_function(float x)
+double act_squashing_function(double x)
 {
 	return (1/(1+(exp(-x))));
 }
 
-float act_step_function(float x)
+double act_step_function(double x)
 {
 	if(0 <= x)
 		return 1;
